Fixed up factorial.
Before, if I entered a huge number that overflowed the "int" input to
the factorial function it would return unexpected results.

+++ include/factorial.h
@@ -1,6 +1,6 @@
 #ifndef _FACTORIAL_
 #define _FACTORIAL_
 
-long double factorial(int x);
+long double factorial(long double x);
 
-#endif
+#endif // _FACTORIAL_
